Resolve "fix user delete checking current user"

Closes #69

See merge request jurvanen/api-access-control!51

+++ backend/src/schema/resolvers/mutation_tyk_deleteUser.ts
@@ -6,7 +6,7 @@ import {
 import { prisma } from "../../generated/prisma-client";
 import logger from "../../misc/logger";
 import fetch from "node-fetch";
-import { DEBUG, TYK_GW_SECRET } from "../../environment";
+import { DEBUG, TYK_GW_SECRET, ROOT_ADMIN_EMAIL } from "../../environment";
 
 export default {
   Mutation: {
@@ -35,21 +35,26 @@ export default {
       for (var i = 0; i < input.user_ids.length; i++) {
         const userTemp = await prisma.user({ id: input.user_ids[i] });
 
-        const baseUrl = "http://gateway:8080/tyk/";
-        const keysPath = "keys/";
+        if (
+          userTemp.email !== ROOT_ADMIN_EMAIL &&
+          userTemp.id !== currentUser.id
+        ) {
+          const baseUrl = "http://gateway:8080/tyk/";
+          const keysPath = "keys/";
 
-        const headers = {
-          "Content-Type": "application/json",
-          "x-tyk-authorization": TYK_GW_SECRET
-        };
+          const headers = {
+            "Content-Type": "application/json",
+            "x-tyk-authorization": TYK_GW_SECRET
+          };
 
-        const url = baseUrl + keysPath + userTemp.api_hash + "?hashed=true";
+          const url = baseUrl + keysPath + userTemp.api_hash + "?hashed=true";
 
-        await fetch(url, { method: "DELETE", headers: headers });
+          await fetch(url, { method: "DELETE", headers: headers });
 
-        const u = await prisma.deleteUser({ id: input.user_ids[i] });
+          const u = await prisma.deleteUser({ id: input.user_ids[i] });
 
-        user.push(u);
+          user.push(u);
+        }
       }
 
       return { user };
